# api-spec.yaml
swagger: "2.0"
info:
  title: Signature API Gateway
  version: "1.0.0"
schemes:
  - https
produces:
  - application/json
x-google-backend:
  address: https://signature-api-yareqqxdwq-du.a.run.app
  deadline: 30.0
security:
  - api_key: []
securityDefinitions:
  api_key:
    type: apiKey
    name: x-api-key
    in: header
paths:
  /:
    get:
      operationId: root
      responses:
        200:
          description: OK
  /health:
    get:
      operationId: health
      responses:
        200:
          description: OK

  # Sales API
  /api/sales:
    post:
      operationId: createSale
      responses:
        201:
          description: Created
  /api/sales/{id}:
    get:
      operationId: getSale
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
  /api/sales/{id}/refund:
    put:
      operationId: refundSale
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK

  # Settlement API
  /api/settlement/monthly:
    get:
      operationId: monthlySettlement
      responses:
        200:
          description: OK
  /api/settlement/domestic:
    get:
      operationId: domesticSettlement
      responses:
        200:
          description: OK
  /api/settlement/overseas:
    get:
      operationId: overseasSettlement
      responses:
        200:
          description: OK

  # Exchange Rates API
  /api/exchange-rates:
    get:
      operationId: getLatestExchangeRates
      responses:
        200:
          description: OK
  /api/exchange-rates/{date}:
    get:
      operationId: getExchangeRatesByDate
      parameters:
        - name: date
          in: path
          required: true
          type: string
          format: date
      responses:
        200:
          description: OK

  # Popups API
  /api/popups:
    get:
      operationId: getPopups
      responses:
        200:
          description: OK
    post:
      operationId: createPopup
      responses:
        201:
          description: Created
  /api/popups/active:
    get:
      operationId: getActivePopups
      responses:
        200:
          description: OK
  /api/popups/{id}:
    get:
      operationId: getPopupById
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
    put:
      operationId: updatePopup
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
    delete:
      operationId: deletePopup
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
  /api/popups/{id}/status:
    put:
      operationId: updatePopupStatus
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK

  # Stores API
  /api/stores:
    get:
      operationId: getStores
      responses:
        200:
          description: OK
    post:
      operationId: createStore
      responses:
        201:
          description: Created
  /api/stores/{id}:
    get:
      operationId: getStoreById
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
    put:
      operationId: updateStore
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
    delete:
      operationId: deleteStore
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK

  # Devices API
  /api/devices:
    get:
      operationId: getDevices
      responses:
        200:
          description: OK
    post:
      operationId: createDevice
      responses:
        201:
          description: Created
  /api/devices/{id}:
    get:
      operationId: getDeviceById
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
    put:
      operationId: updateDevice
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
    delete:
      operationId: deleteDevice
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK

  # Config API
  /api/config:
    get:
      operationId: getAllConfig
      responses:
        200:
          description: OK
  /api/config/{key}:
    get:
      operationId: getConfigByKey
      parameters:
        - name: key
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
    put:
      operationId: updateConfig
      parameters:
        - name: key
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK